ai/embeddings/requirements.txt
redis
redis[hiredis]
openai
pymongo
pymongo[srv]
python-dotenv
numpy